+++ deepr/layers/lookup.py
@@ -3,18 +3,17 @@
 from typing import Dict, Callable
 import logging
 
-import numpy as np
 import tensorflow as tf
 
 from deepr.layers import base
-from deepr.utils.field import TensorType
+from deepr.utils.tables import table_from_mapping, table_from_file, index_to_string_table_from_file
 
 
 LOGGER = logging.getLogger(__name__)
 
 
 class Lookup(base.Layer):
-    """Lookup Layer
+    """Lookup Layer.
 
     Attributes
     ----------
@@ -32,7 +31,7 @@ class Lookup(base.Layer):
 
 
 class LookupFromFile(Lookup):
-    """Lookup From File Layer
+    """Lookup From File Layer.
 
     Creates a table at runtime from a mapping file. The table will map
     each key to its corresponding line index as an tf.int64.
@@ -43,19 +42,24 @@ class LookupFromFile(Lookup):
         Keys type
     path : str
         Path to mapping file
+    reuse : bool
+        If True, reuse table with the same name
     table_name : str
         Name of the HashTable
     """
 
-    def __init__(self, table_name: str, path: str, key_dtype=None, **kwargs):
-        super().__init__(lambda: table_from_file(name=table_name, path=path, key_dtype=key_dtype), **kwargs)
+    def __init__(self, table_name: str, path: str, key_dtype=None, reuse: bool = False, **kwargs):
+        super().__init__(
+            lambda: table_from_file(name=table_name, path=path, key_dtype=key_dtype, reuse=reuse), **kwargs
+        )
         self.table_name = table_name
         self.path = path
         self.key_dtype = key_dtype
+        self.reuse = reuse
 
 
 class LookupFromMapping(Lookup):
-    """Lookup From Mapping Layer
+    """Lookup From Mapping Layer.
 
     Attributes
     ----------
@@ -65,13 +69,24 @@ class LookupFromMapping(Lookup):
         Keys type
     mapping : Dict[Any, Any]
         Mapping keys -> index
+    reuse : bool
+        If True, reuse the layer with the same name
     table_name : str
         Name of the HashTable
     value_dtype : tf.DType
         Values type
     """
 
-    def __init__(self, table_name: str, mapping: Dict, default_value=None, key_dtype=None, value_dtype=None, **kwargs):
+    def __init__(
+        self,
+        table_name: str,
+        mapping: Dict,
+        default_value=None,
+        key_dtype=None,
+        value_dtype=None,
+        reuse: bool = False,
+        **kwargs,
+    ):
         super().__init__(
             lambda: table_from_mapping(
                 name=table_name,
@@ -79,6 +94,7 @@ class LookupFromMapping(Lookup):
                 default_value=default_value,
                 key_dtype=key_dtype,
                 value_dtype=value_dtype,
+                reuse=reuse,
             ),
             **kwargs,
         )
@@ -87,39 +103,46 @@ class LookupFromMapping(Lookup):
         self.default_value = default_value
         self.key_dtype = key_dtype
         self.value_dtype = value_dtype
+        self.reuse = reuse
 
 
-def table_from_file(name: str, path: str, key_dtype=None):
-    """Create table from file"""
-    LOGGER.info(f"Creating table {name} from {path}")
-    table = tf.contrib.lookup.index_table_from_file(vocabulary_file=path, name=name, key_dtype=key_dtype)
-    return table
+class LookupIndexToString(Lookup):
+    """Lookup Index To String.
 
+    Creates a table at runtime from a mapping file. The table will map
+    each key to its corresponding line index as an tf.int64.
 
-def table_from_mapping(name: str, mapping: Dict, default_value=None, key_dtype=None, value_dtype=None):
-    """Create table from mapping"""
-    LOGGER.info(f"Creating table {name} from mapping.")
+    Attributes
+    ----------
+    default_value : Any
+        Default Value for missing keys
+    path : str
+        Path to mapping file
+    reuse : bool
+        If True, reuse the table with the same name
+    table_name : str
+        Name of the HashTable
+    vocab_size : int
+        Size of the vocab
+    """
 
-    # Convert mapping to arrays of keys and values
-    keys, values = zip(*mapping.items())  # type: ignore
-    keys_np = np.array(keys)
-    values_np = np.array(values)
-
-    # Infer default value if not given
-    if default_value is None:
-        default_value = TensorType(type(values_np[0].item())).default
-
-    # Infer types if not given
-    if key_dtype is None:
-        key_dtype = TensorType(type(keys_np[0].item())).tf
-    if value_dtype is None:
-        value_dtype = TensorType(type(values_np[0].item())).tf
-
-    # Create table
-    return tf.contrib.lookup.HashTable(
-        tf.contrib.lookup.KeyValueTensorInitializer(
-            keys=keys_np, values=values_np, key_dtype=key_dtype, value_dtype=value_dtype
-        ),
-        name=name,
-        default_value=default_value,
-    )
+    def __init__(
+        self,
+        table_name: str,
+        path: str = None,
+        vocab_size: int = None,
+        default_value="UNK",
+        reuse: bool = False,
+        **kwargs,
+    ):
+        super().__init__(
+            lambda: index_to_string_table_from_file(
+                name=table_name, path=path, vocab_size=vocab_size, default_value=default_value, reuse=reuse
+            ),
+            **kwargs,
+        )
+        self.table_name = table_name
+        self.path = path
+        self.vocab_size = vocab_size
+        self.default_value = default_value
+        self.reuse = reuse