+++ deepr/vocab/__init__.py
@@ -1,0 +1,3 @@
+# pylint: disable=unused-import,missing-docstring
+
+from deepr.vocab.base import size, read, write