+++ tests/unit/prepros/test_prepros_base.py
@@ -1,4 +1,4 @@
-# pylint: disable=no-value-for-parameter,invalid-name,unexpected-keyword-arg
+# pylint: disable=no-value-for-parameter,invalid-name,unexpected-keyword-arg,redefined-outer-name
 """Tests for prepros.base"""
 
 import pytest
@@ -8,138 +8,97 @@ import numpy as np
 import deepr as dpr
 
 
-def gen():
-    yield {"a": [0]}
-    yield {"a": [0, 1]}
+@pytest.fixture
+def dataset():
+    def _gen():
+        yield {"a": [0]}
+        yield {"a": [0, 1]}
+
+    return tf.data.Dataset.from_generator(_gen, {"a": tf.int32}, {"a": (None,)})
 
 
-def test_prepros_decorator_from_apply():
+@dpr.prepros.prepro
+def AddOffset(dataset, offset):
+    """AddOffset"""
+    return dataset.map(lambda x: {"b": x["a"] + offset})
+
+
+@dpr.prepros.prepro
+def AddOne(dataset):
+    """AddOne"""
+    return dataset.map(lambda x: {"b": x["a"] + 1})
+
+
+def test_prepros_decorator_from_apply(dataset):
     """Create preprocessor from an apply function"""
+    # Check decorated function properties
+    assert issubclass(AddOffset, dpr.prepros.Prepro)
+    assert AddOffset.__name__ == "AddOffset"
+    assert AddOffset.__doc__ == "AddOffset"
+    assert AddOffset.__module__ == __name__
 
-    @dpr.prepros.prepro
-    def AddOffset(dataset, offset):
-        return dataset.map(lambda x: {"b": x["a"] + offset})
-
-    ds = tf.data.Dataset.from_generator(gen, {"a": tf.int32}, {"a": (None,)})
-
-    # Positional argument
-    add_one = AddOffset(1)
-    reader = dpr.readers.from_dataset(add_one(ds))
-    expected = [{"b": [1]}, {"b": [1, 2]}]
-    np.testing.assert_equal(list(reader), expected)
-
-    # Keyword argument
+    # Check instance properties
     add_one = AddOffset(offset=1)
-    reader = dpr.readers.from_dataset(add_one(ds))
+    reader = dpr.readers.from_dataset(add_one(dataset))
     expected = [{"b": [1]}, {"b": [1, 2]}]
     np.testing.assert_equal(list(reader), expected)
 
 
-def test_prepros_decorator_from_constructor():
-    """Create preprocessor from a Prepro constructor"""
+def test_prepros_decorator_from_apply_laziness(dataset):
+    """Laziness is especially useful if custom prepro use hash tables"""
 
     @dpr.prepros.prepro
-    def AddOffset(offset):
-        return dpr.prepros.Map(lambda x: {"b": x["a"] + offset}, update=False)
+    def RaiseApply(dataset):
+        raise RuntimeError()
 
-    ds = tf.data.Dataset.from_generator(gen, {"a": tf.int32}, {"a": (None,)})
-
-    # Positional argument
-    add_one = AddOffset(1)
-    reader = dpr.readers.from_dataset(add_one(ds))
-    expected = [{"b": [1]}, {"b": [1, 2]}]
-    np.testing.assert_equal(list(reader), expected)
-
-    # Keyword argument
-    add_one = AddOffset(offset=1)
-    reader = dpr.readers.from_dataset(add_one(ds))
-    expected = [{"b": [1]}, {"b": [1, 2]}]
-    np.testing.assert_equal(list(reader), expected)
+    prepro = RaiseApply()
+    with pytest.raises(RuntimeError):
+        prepro(dataset)
 
 
-def test_prepros_decorator_signatures():
-    """Test incorrect use of decorator"""
-    # pylint: disable=unused-argument,unused-variable,too-many-function-args
-
-    # Simple preprocessor with no arguments
-    @dpr.prepros.prepro
-    def Simple(dataset) -> tf.data.Dataset:
-        pass
-
-    Simple()
-
-    with pytest.raises(TypeError):
-        Simple(dataset=1)
-
-    with pytest.raises(TypeError):
-        Simple(1)
-
-    with pytest.raises(TypeError):
-        Simple(foo=1)
-
-    # Typical preprocessor with positional and keyword arguments
-    @dpr.prepros.prepro
-    def Typical(dataset, mode, foo, bar=1) -> tf.data.Dataset:
-        pass
-
-    Typical(1)
-    Typical(1, 2)
-    Typical(foo=1, bar=2)
-    Typical(1, bar=2)
-
-    with pytest.raises(TypeError):
-        Typical(1, 2, 3)
-
-    with pytest.raises(TypeError):
-        Typical(1, foo=1)
-
-    with pytest.raises(TypeError):
-        Typical(1, baz=1)
-
-    # From constructor
-    def FromConstructor(foo, bar=1) -> dpr.prepros.Prepro:
-        pass
-
-    FromConstructor(1, 2)
-    FromConstructor(1, bar=2)
-    FromConstructor(foo=1, bar=2)
-
-    with pytest.raises(TypeError):
-        FromConstructor(1, 2, 3)
-
-    with pytest.raises(TypeError):
-        Typical(1, foo=1)
-
-    with pytest.raises(TypeError):
-        Typical(1, baz=1)
-
-    # Test wrong order in arguments raises error at decoration time
+def test_prepros_decorator_from_apply_wrong_order():
+    """Test wrong order in arguments raises error at decoration time."""
+    # pylint: disable=unused-variable
     with pytest.raises(TypeError):
 
         @dpr.prepros.prepro
         def WrongOrder(offset, dataset):
-            pass
+            return dataset.map(lambda x: {"b": x["a"] + offset})
 
 
-def test_prepros_decorator_laziness_from_apply():
-    """Laziness is especially useful if custom prepro use hash tables"""
-
-    @dpr.prepros.prepro
-    def ErrorPrepro(dataset):
-        raise ValueError()
-
-    prepro_fn = ErrorPrepro()
-    with pytest.raises(ValueError):
-        prepro_fn(None)
+@dpr.prepros.prepro
+def Identity(dataset) -> tf.data.Dataset:
+    return dataset
 
 
-def test_prepros_decorator_laziness_from_constructor():
-    """Laziness is especially useful if custom prepro use hash tables"""
+@dpr.prepros.prepro
+def Typical(dataset, foo, bar=1) -> dpr.prepros.Prepro:
+    return dataset.map(lambda x: {"b": x["a"] + foo + 2 * bar})
 
-    @dpr.prepros.prepro
-    def ErrorPrepro():
-        raise ValueError()
 
-    prepro_fn = ErrorPrepro()
-    with pytest.raises(ValueError):
-        prepro_fn(None)
+@pytest.mark.parametrize(
+    "cls, args, kwargs, error",
+    [
+        # Simple layer with no special arguments
+        (Identity, (), {}, None),
+        (Identity, (), {"tensors": 1}, TypeError),
+        (Identity, (1,), None, TypeError),
+        (Identity, (), {"foo": 1}, TypeError),
+        # Typical layers with positional and keyword arguments
+        (Typical, (1,), {}, None),
+        (Typical, (1, 2), {}, None),
+        (Typical, (), {"foo": 1, "bar": 2}, None),
+        (Typical, (1,), {"bar": 2}, None),
+        (Typical, (1, 2, 3), {}, TypeError),
+        (Typical, (1,), {"foo": 1}, TypeError),
+        (Typical, (1,), {"baz": 1}, TypeError),
+    ],
+)
+def test_prepros_decorator_instantiation(cls, args, kwargs, error):
+    """Test prepro instantiation from decorator."""
+    if error is not None:
+        with pytest.raises(error):
+            cls(*args, **kwargs)
+    else:
+        instance = cls(*args, **kwargs)
+        assert isinstance(instance, dpr.prepros.Prepro)